pyproject.toml
[project]
name = "codebook-archive"
version = "0.1.0"
description = "Discover, classify, and host political-communication codebooks from OA sources."
authors = [{ name = "Eva Vogel", email = "eva_vogel@web.de" }]
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "typer>=0.12",
    "rich>=13.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
extract = ["pypdf>=4.0", "pymupdf>=1.24"]
classify = ["anthropic>=0.40"]
review = ["streamlit>=1.30", "pandas>=2.0"]
site = ["mkdocs-material>=9.5", "pagefind>=1.0"]

[project.scripts]
codebook-archive = "codebook_archive.__main__:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/codebook_archive"]
